* commit '4e17065ef48ba8a774f415ba829a6b111d7dc95a': (23 commits)
  Use constants instead of unnamed enum.
  Update b2code-cppcheck to C++20
  Fix other few -Wdeprecated-enum-float-conversion warnings in vxd package
  Fix few -Wdeprecated-enum-float-conversion and -Wdeprecated-anon-enum-enum-conversion warnings in tracking package
  Fix other few -Wdeprecated-enum-float-conversion warnings in top package
  Fix a -Wdeprecated-enum-float-conversion warning in BKLMSimulationPar class
  Fix a -Wdeprecated-enum-float-conversion warning in ECLChannelMapper class
  Fix few -Wdeprecated-anon-enum-enum-conversion warnings in analysis package
  Fix other few -Wdeprecated-anon-enum-enum-conversion warnings in rawdata package
  Fix a -Wunused-variable warning in CDCTriggerNeuro module
  Partial revert commit 748873f5d53: apply only minimal modifications to rawdata package for fixing the compiler warnings
  Fix deprecated implicit capture of this via [=] in XHitFilterFactory lambdas, since it is deprecated in C++20
  Fix few -Wdeprecated-anon-enum-enum-conversion warnings in ecl package
  Silence a -Wtype-limits warning in DQMHistAnalysisInputPVSrv module
  Fix few -Wdeprecated-anon-enum-enum-conversion warnings in rawdata package
  Fix few -Wdeprecated-enum-float-conversion warnings in top package
  Fix the -Wambiguous-reversed-operator warning in KLMChannelIndex class
  Revert modifications to ROIid class; Modify the getters/setters for VxdId accordingly
  Revert "Make GenFit compiling with C++20"
  Fix compilation with gcc/g++
  ...

+++ top/modules/TOPBunchFinder/src/TOPBunchFinderModule.cc
@@ -241,7 +241,7 @@ namespace Belle2 {
     // full time window in which data are taken (smaller time window is used in reconstruction)
 
     const auto& tdc = TOPGeometryPar::Instance()->getGeometry()->getNominalTDC();
-    double timeWindow = m_feSetting->getReadoutWindows() * tdc.getSyncTimeBase() / TOPNominalTDC::c_syncWindows;
+    double timeWindow = m_feSetting->getReadoutWindows() * tdc.getSyncTimeBase() / static_cast<double>(TOPNominalTDC::c_syncWindows);
 
     // counters and temporary containers
